inventario_web/inventario/requirements.txt
mysqlclient
django-admin-interface
django-crispy-forms
crispy-bootstrap4